Rename `FeedViewControllerTests` to `FeedUIIntegrationTests` since we are now testing the composition of multiple UI components in integration. Extract helper methods in separate files and group them together under a Helpers folder
