+++ EssentialFeediOSTests/Feed UI/Helpers/FeedImageCell+TestHelpers.swift
@@ -1,0 +1,41 @@
+//
+//  FeedImageCell+TestHelpers.swift
+//  EssentialFeediOSTests
+//
+//  Created by TCode on 12/5/22.
+//
+
+import Foundation
+import XCTest
+import EssentialFeediOS
+
+extension FeedImageCell {
+    func simulateRetryAction() {
+        feedImageRetryButton.simulateTap()
+    }
+
+    var isShowingLocation: Bool {
+        return !locationContainer.isHidden
+    }
+
+    var isShowingImageLoadingIndicator: Bool {
+        return feedImageContainer.isShimmering
+    }
+
+    var isShowingRetryAction: Bool {
+        return !feedImageRetryButton.isHidden
+    }
+
+    var locationText: String? {
+        return locationLabel.text
+    }
+
+    var descriptionText: String? {
+        return descriptionLabel.text
+    }
+
+    var renderedImage: Data? {
+        return feedImageView.image?.pngData()
+    }
+}
+
