+++ EssentialFeediOSTests/Feed UI/Helpers/UIRefreshControl+TestHelpers.swift
@@ -1,0 +1,17 @@
+//
+//  UIRefreshControl+TestHelpers.swift
+//  EssentialFeediOSTests
+//
+//  Created by TCode on 12/5/22.
+//
+
+import UIKit
+
+extension UIRefreshControl {
+    func simulatePullToRefresh() {
+        allTargets.forEach({ target in
+            actions(forTarget: target, forControlEvent: .valueChanged)?
+                .forEach { (target as NSObject).perform(Selector($0)) }
+        })
+    }
+}
